refactor: split SimpleLocale into Localize.Runtime and Localize.Routes
