+++ lib/routex/extension/runtime_callbacks.ex
@@ -12,8 +12,9 @@ defmodule Routex.Extension.RuntimeCallbacks do
   > #### In combination with... {: .neutral}
   > This extension calls other functions with values from `Routex.Attrs` during
   > runtime. These attributes can be set by other extensions such as
-  > `Routex.Extension.Alternatives` (compile time) and
-  > `Routex.Extension.SimpleLocale` (run time)
+  > `Routex.Extension.Alternatives` (compile time),
+  > `Routex.Extension.Localize.Routes` (compile time)
+  > and `Routex.Extension.Localize.Runtime` (run time)
 
 
   ### Options
